introduce delete endpoint
